requirements.txt
pytelegrambotapi==4.16.0
python-dotenv==1.0.0
jinja2==3.1.2
prometheus-client==0.17
